requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
wikipedia>=1.4.0
python-dateutil>=2.8.0
black>=23.0.0
flake8>=6.0.0